+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 **Features**:
 
 * Integrate Web Console, [PR-77](https://github.com/reduct-storage/reduct-storage/pull/77)
+* Print only HTTP errors 50x to the logs, [PR-84](https://github.com/reduct-storage/reduct-storage/issues/84)
 
 **Bugs**: