+++ CHANGELOG.md
@@ -11,6 +11,10 @@
 * Fix CORS functionality, [PR-72](https://github.com/reduct-storage/reduct-storage/pull/72)
 * Fix quota policy, [PR-83](https://github.com/reduct-storage/reduct-storage/pull/83)
 
+**Documentation**:
+
+* Add link to JS SDK, [PR-88](https://github.com/reduct-storage/reduct-storage/pull/88)
+
 ### Release 0.4.3 (2022-05-01)
 
 **Bugs**: